API works properly

+++ src/main/java/com/svalero/RosasTattoo/domain/Client.java
@@ -42,5 +42,5 @@ public class Client {
     private boolean showPhoto;
 
     @Column
-    private long visits;
+    private int visits;
 }
